+++ integration-tests/src/main/java/com/kalixia/grapi/EchoResource.java
@@ -1,5 +1,6 @@
 package com.kalixia.grapi;
 
+import javax.inject.Inject;
 import javax.validation.constraints.NotNull;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
@@ -10,6 +11,10 @@ import javax.ws.rs.core.MediaType;
 @Path("/echo")
 @Produces(MediaType.TEXT_PLAIN)
 public class EchoResource {
+
+    @Inject
+    public EchoResource() {
+    }
 
     @Path("{message}")
     @GET