+++ integration-tests/src/main/java/com/kalixia/grapi/HeadersResource.java
@@ -1,11 +1,16 @@
 package com.kalixia.grapi;
 
+import javax.inject.Inject;
 import javax.ws.rs.GET;
 import javax.ws.rs.HeaderParam;
 import javax.ws.rs.Path;
 
 @Path("/headers")
 public class HeadersResource {
+
+    @Inject
+    public HeadersResource() {
+    }
 
     @GET
     public String echoHeader(@HeaderParam("X-CUSTOM-HEADER") String header) {