+++ integration-tests/src/main/java/com/kalixia/grapi/HelloResource.java
@@ -1,11 +1,16 @@
 package com.kalixia.grapi;
 
+import javax.inject.Inject;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 
 @Path("/hello")
 public class HelloResource {
+
+    @Inject
+    public HelloResource() {
+    }
 
     @GET
     public String hello() {
@@ -14,7 +19,7 @@ public class HelloResource {
 
     @GET
     @Path("{lastName}/{firstName}")
-    public String hello2(@PathParam("firstName") String first, @PathParam("lastName") String last) {
+    public String helloFromFirstNameAndLastName(@PathParam("firstName") String first, @PathParam("lastName") String last) {
         return String.format("Hello %s %s!", first, last);
     }