+++ integration-tests/src/main/java/com/kalixia/grapi/QueryParamResource.java
@@ -1,5 +1,6 @@
 package com.kalixia.grapi;
 
+import javax.inject.Inject;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.QueryParam;
@@ -7,20 +8,24 @@ import javax.ws.rs.QueryParam;
 @Path("/query")
 public class QueryParamResource {
 
+    @Inject
+    public QueryParamResource() {
+    }
+
     @GET
-    String echoQueryParam(@QueryParam("message") String msg) {
+    public String echoQueryParam(@QueryParam("message") String msg) {
         return msg;
     }
 
     @GET
     @Path("/number")
-    String echoNumber(@QueryParam("number") Integer number) {
+    public String echoNumber(@QueryParam("number") Integer number) {
         return number.toString();
     }
 
     @GET
     @Path("/number-primitive")
-    String echoNumberPrimitive(@QueryParam("number") int number) {
+    public String echoNumberPrimitive(@QueryParam("number") int number) {
         return Integer.toString(number);
     }