+++ integration-tests/src/main/java/com/kalixia/grapi/TestModule.java
@@ -1,7 +1,0 @@
-package com.kalixia.grapi;
-
-import dagger.Module;
-
-@Module(injects = TestDagger.class, includes = GeneratedJaxRsDaggerModule.class)
-class TestModule {
-}