+++ netty-compiler/src/main/java/com/kalixia/grapi/apt/jaxrs/model/JaxRsMethodInfoComparator.java
@@ -13,6 +13,7 @@ public class JaxRsMethodInfoComparator implements Comparator<JaxRsMethodInfo>, S
         // compare verbs first
         HttpMethod verb1 = HttpMethod.valueOf(method1.getVerb());
         HttpMethod verb2 = HttpMethod.valueOf(method2.getVerb());
+        // compare uri templates next
         return verb1.equals(verb2) ?
                 templateComparator.compare(method1.getUriTemplate(), method2.getUriTemplate())
                 : verb1.compareTo(verb2);